pyproject.toml
[tool.poetry]
name = "ygg-rss-proxy"
version = "1.0.0-alpha"
description = "Proxy for yggtorrent rss feeds"
authors = ["LimeDrive <limecat@limedrive.eu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
requests = "^2.32.3"
lxml = "^5.2.2"
gunicorn = "^22.0.0"
pydantic = "^2.7.3"
loguru = "^0.7.2"
pydantic-settings = "^2.3.0"
flask-session = "^0.8.0"
flask-sqlalchemy = "^3.1.1"
orjson = "^3.10.4"
stackprinter = "^0.2.12"
timeout-decorator = "^0.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
